feat(terminal): factored out components, made improvements
