+++ src/app/globals.css
@@ -83,7 +83,7 @@
 	}
 }
 
-.type::after {
+.type:focus-within::after {
 	content: '|';
 	animation: cursor 1.1s infinite step-start;
 }
