+++ src/components/visual-terminal/visual-terminal.tsx
@@ -10,7 +10,7 @@ type Props = PropsWithChildren & {
 export const VisualTerminal = ({ children, title, onClick, className }: Props) => {
 	return (
 		<div className={cn('pt-5', className)} onClick={onClick}>
-			<div className="mt-15 mx-auto w-full max-w-3xl rounded-lg bg-terminal-body pb-4 text-white shadow-md">
+			<div className="mt-15 mx-auto w-full max-w-3xl rounded-lg border-[1px] border-white border-opacity-20 bg-terminal-body pb-4 text-white shadow-md">
 				<div className="flex items-center justify-between rounded-t-lg bg-terminal-header p-2">
 					<div className="flex items-center space-x-2 text-xs">
 						<div className="h-3 w-3 rounded-full bg-red-500"></div>
