+++ tailwind.config.js
@@ -38,8 +38,7 @@ module.exports = {
 				background: 'hsl(var(--background))',
 				foreground: 'hsl(var(--foreground))',
 				'terminal-header': '#333333',
-				'terminal-body': '#0c0c0c',
-				'terminal-alt': '#151718',
+				'terminal-body': '#1e1e1e',
 				primary: {
 					DEFAULT: 'hsl(var(--primary))',
 					foreground: 'hsl(var(--primary-foreground))',
